refactor(FON-468): flatten SG page folders, extract presentation steps, rename member space
