+++ apps/client/src/pages/member/SessionsPage.tsx
