+++ apps/client/src/pages/secretariat-general/DetailsMemberPage.tsx
@@ -1,9 +1,9 @@
 import { Navigate, useParams } from 'react-router';
 
-import { PageContentLayout } from '../../../components/shared/PageContentLayout';
-import { ROUTE_PATHS } from '../../../utils/route-path.utils';
+import { PageContentLayout } from '@/components/shared/PageContentLayout';
 import { DetailsMember } from '@/features/members/components/DetailsMember';
 import { DetailsMemberHeader } from '@/features/members/components/DetailsMemberHeader';
+import { ROUTE_PATHS } from '@/utils/route-path.utils';
 import { useDetailedMember } from '@queries/members.queries';
 
 export function DetailsMemberPage() {
