+++ apps/client/src/pages/secretariat-general/ObservationDetailsPage.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Navigate, useParams, useSearchParams } from 'react-router';
 
-import { ObservationDetailsContent } from '../../../components/shared/ObservationDetailsContent';
-import { PageContentLayout } from '../../../components/shared/PageContentLayout';
-import { getDetailSessionGdsPath, ROUTE_PATHS } from '../../../utils/route-path.utils';
 import { ArchiveBannerPortal } from '@/components/shared/layouts/archived-banner/ArchiveBannerPortal';
+import { ObservationDetailsContent } from '@/components/shared/ObservationDetailsContent';
 import { ObservationFollowUpCommentProvider } from '@/components/shared/observations/follow-up-selector/ObservationFollowUpCommentDialogProvider';
+import { PageContentLayout } from '@/components/shared/PageContentLayout';
 import type { FilesUploader } from '@/components/shared/ui/tip-tap-editor/extensions/editor-file-uploader';
 import { useIsSgNavigation } from '@/hooks/roles.hook';
+import { getDetailSessionGdsPath, ROUTE_PATHS } from '@/utils/route-path.utils';
 import {
   useAttachObservationMemberCommentScreenshotsMutation,
   useGetObservationFileUrlMutation,
